move eraser menu location
move menu to picHub, minimized execution time
