+++ bin/.gitignore
@@ -1,0 +1,5 @@
+/Test$1.class
+/Test.class
+/PictHub$6.class
+/PictHub$7.class
+/PictHub$8.class
